Moved the next, prev arrows into the 'icons' folder. Update they import
